backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
openai==1.57.2
pydantic==2.10.3
python-multipart==0.0.17
pymupdf==1.25.1
python-dotenv==1.0.1
google-generativeai>=0.8.0
instructor>=1.0.0
marker-pdf>=1.0.0
